cinema_together/requirements.txt
fastapi==0.109.0
alembic==1.13.1
psycopg2==2.9.6
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
uvicorn==0.27.0
asyncpg==0.29.0
orjson==3.8.10
pyjwt==2.8.0
aiohttp==3.9.1
